+++ game_patch/hud/hud.cpp
@@ -446,4 +446,5 @@ void hud_apply_patches()
     multi_hud_apply_patches();
     message_log_apply_patch();
     hud_world_apply_patch();
+    hud_colors_apply_patch();
 }