+++ game_patch/misc/alpine_settings.cpp
@@ -41,6 +41,81 @@ std::optional<std::string> afs_cmd_line_filename;
 
 AlpineGameSettings g_alpine_game_config;
 
+std::optional<uint32_t> parse_hex_color_string(const std::string& value)
+{
+    std::string sanitized;
+    sanitized.reserve(value.size());
+
+    // Remove whitespace
+    for (unsigned char c : value) {
+        if (!std::isspace(c)) {
+            sanitized.push_back(static_cast<char>(c));
+        }
+    }
+
+    // Remove 0x prefix
+    if (string_starts_with_ignore_case(sanitized, "0x")) {
+        sanitized.erase(0, 2);
+    }
+
+    // Validate length
+    if (sanitized.size() != 6 && sanitized.size() != 8) {
+        return std::nullopt;
+    }
+
+    // Validate format
+    if (!std::all_of(sanitized.begin(), sanitized.end(), [](unsigned char c) { return std::isxdigit(c) != 0; })) {
+        return std::nullopt;
+    }
+
+    // Set color, assuming fully opaque if alpha value is not specified
+    uint32_t color = 0;
+    auto result = std::from_chars(sanitized.data(), sanitized.data() + sanitized.size(), color, 16);
+    if (result.ec != std::errc{}) {
+        return std::nullopt;
+    }
+
+    if (sanitized.size() == 6) {
+        color = (color << 8) | 0xFF;
+    }
+
+    return color;
+}
+
+std::string format_hex_color_string(uint32_t color)
+{
+    std::ostringstream stream;
+    uint8_t alpha = static_cast<uint8_t>(color & 0xFF);
+    bool include_alpha = alpha != 0xFF;
+    uint32_t encoded_value = include_alpha ? color : (color >> 8);
+
+    stream << std::uppercase << std::hex << std::setfill('0')
+           << std::setw(include_alpha ? 8 : 6) << encoded_value;
+    return stream.str();
+}
+
+// parse colors to 0-255 ints
+std::tuple<int, int, int, int> extract_color_components(uint32_t color)
+{
+    return std::make_tuple(
+        (color >> 24) & 0xFF, // Red
+        (color >> 16) & 0xFF, // Green
+        (color >> 8) & 0xFF,  // Blue
+        color & 0xFF          // Alpha
+    );
+}
+
+// parse colors to 0.0-1.0 floats
+std::tuple<float, float, float, float> extract_normalized_color_components(uint32_t color)
+{
+    return std::make_tuple(
+        ((color >> 24) & 0xFF) / 255.0f, // Red   (0-1)
+        ((color >> 16) & 0xFF) / 255.0f, // Green (0-1)
+        ((color >> 8)  & 0xFF) / 255.0f, // Blue  (0-1)
+        (color & 0xFF) / 255.0f          // Alpha (0-1)
+    );
+}
+
 static rf::CmdLineParam& get_afs_cmd_line_param()
 {
     static rf::CmdLineParam afs_param{"-afs", "", true};
@@ -323,11 +398,6 @@ bool alpine_player_settings_load(rf::Player* player)
         g_alpine_game_config.set_fpgun_fov_scale(std::stof(settings["FPGunFOVScale"]));
         processed_keys.insert("FPGunFOVScale");
     }
-    if (settings.count("BigHUD")) {
-        g_alpine_game_config.big_hud = std::stoi(settings["BigHUD"]);
-        set_big_hud(g_alpine_game_config.big_hud);
-        processed_keys.insert("BigHUD");
-    }
     if (settings.count("DisableWeaponShake")) {
         g_alpine_game_config.try_disable_weapon_shake = std::stoi(settings["DisableWeaponShake"]);
         processed_keys.insert("DisableWeaponShake");
@@ -343,10 +413,6 @@ bool alpine_player_settings_load(rf::Player* player)
     if (settings.count("DisableMuzzleFlashLights")) {
         g_alpine_game_config.try_disable_muzzle_flash_lights = std::stoi(settings["DisableMuzzleFlashLights"]);
         processed_keys.insert("DisableMuzzleFlashLights");
-    }
-    if (settings.count("ReticleScale")) {
-        g_alpine_game_config.set_reticle_scale(std::stof(settings["ReticleScale"]));
-        processed_keys.insert("ReticleScale");
     }
     if (settings.count("ShowGlares")) {
         g_alpine_game_config.show_glares = std::stoi(settings["ShowGlares"]);
@@ -396,6 +462,67 @@ bool alpine_player_settings_load(rf::Player* player)
     if (settings.count("AlwaysClampOfficialLightmaps")) {
         g_alpine_game_config.always_clamp_official_lightmaps = std::stoi(settings["AlwaysClampOfficialLightmaps"]);
         processed_keys.insert("AlwaysClampOfficialLightmaps");
+    }
+
+    // Load UI settings
+    if (settings.count("BigHUD")) {
+        g_alpine_game_config.big_hud = std::stoi(settings["BigHUD"]);
+        set_big_hud(g_alpine_game_config.big_hud);
+        processed_keys.insert("BigHUD");
+    }
+    if (settings.count("ReticleScale")) {
+        g_alpine_game_config.set_reticle_scale(std::stof(settings["ReticleScale"]));
+        processed_keys.insert("ReticleScale");
+    }
+    if (settings.count("SniperScopeColor")) {
+        auto color_override = parse_hex_color_string(settings["SniperScopeColor"]);
+        if (color_override) {
+            g_alpine_game_config.sniper_scope_color_override = color_override;
+        }
+        else {
+            xlog::warn("Invalid sniper scope color override: {}", settings["SniperScopeColor"]);
+        }
+        processed_keys.insert("SniperScopeColor");
+    }
+    if (settings.count("PrecisionScopeColor")) {
+        auto color_override = parse_hex_color_string(settings["PrecisionScopeColor"]);
+        if (color_override) {
+            g_alpine_game_config.precision_scope_color_override = color_override;
+        }
+        else {
+            xlog::warn("Invalid precision scope color override: {}", settings["PrecisionScopeColor"]);
+        }
+        processed_keys.insert("PrecisionScopeColor");
+    }
+    if (settings.count("RailScopeColor")) {
+        auto color_override = parse_hex_color_string(settings["RailScopeColor"]);
+        if (color_override) {
+            g_alpine_game_config.rail_scope_color_override = color_override;
+        }
+        else {
+            xlog::warn("Invalid rail scope color override: {}", settings["RailScopeColor"]);
+        }
+        processed_keys.insert("RailScopeColor");
+    }
+    if (settings.count("ArAmmoColor")) {
+        auto color_override = parse_hex_color_string(settings["ArAmmoColor"]);
+        if (color_override) {
+            g_alpine_game_config.ar_ammo_digit_color_override = color_override;
+        }
+        else {
+            xlog::warn("Invalid AR ammo digit color override: {}", settings["ArAmmoColor"]);
+        }
+        processed_keys.insert("ArAmmoColor");
+    }
+    if (settings.count("DamageNotifyColor")) {
+        auto color_override = parse_hex_color_string(settings["DamageNotifyColor"]);
+        if (color_override) {
+            g_alpine_game_config.damage_notify_color_override = color_override;
+        }
+        else {
+            xlog::warn("Invalid damage notification color override: {}", settings["DamageNotifyColor"]);
+        }
+        processed_keys.insert("DamageNotifyColor");
     }
 
     // Load singleplayer settings
@@ -806,12 +933,10 @@ void alpine_player_settings_save(rf::Player* player)
     file << "TextureDetailLevel=" << player->settings.textures_resolution_level << "\n";
     file << "HorizontalFOV=" << g_alpine_game_config.horz_fov << "\n";
     file << "FPGunFOVScale=" << g_alpine_game_config.fpgun_fov_scale << "\n";
-    file << "BigHUD=" << g_alpine_game_config.big_hud << "\n";
     file << "DisableWeaponShake=" << g_alpine_game_config.try_disable_weapon_shake << "\n";
     file << "FullbrightCharacters=" << g_alpine_game_config.try_fullbright_characters << "\n";
     file << "DisableTextures=" << g_alpine_game_config.try_disable_textures << "\n";
     file << "DisableMuzzleFlashLights=" << g_alpine_game_config.try_disable_muzzle_flash_lights << "\n";
-    file << "ReticleScale=" << g_alpine_game_config.reticle_scale << "\n";
     file << "ShowGlares=" << g_alpine_game_config.show_glares << "\n";
     file << "MeshStaticLighting=" << g_alpine_game_config.mesh_static_lighting << "\n";
     file << "Picmip=" << g_alpine_game_config.picmip << "\n";
@@ -823,6 +948,26 @@ void alpine_player_settings_save(rf::Player* player)
     file << "SimulationDistance=" << g_alpine_game_config.entity_sim_distance << "\n";
     file << "FullRangeLighting=" << g_alpine_game_config.full_range_lighting << "\n";
     file << "AlwaysClampOfficialLightmaps=" << g_alpine_game_config.always_clamp_official_lightmaps << "\n";
+
+    // UI
+    file << "\n[UISettings]\n";
+    file << "BigHUD=" << g_alpine_game_config.big_hud << "\n";
+    file << "ReticleScale=" << g_alpine_game_config.reticle_scale << "\n";
+    if (g_alpine_game_config.sniper_scope_color_override) {
+        file << "SniperScopeColor=" << format_hex_color_string(*g_alpine_game_config.sniper_scope_color_override) << "\n";
+    }
+    if (g_alpine_game_config.precision_scope_color_override) {
+        file << "PrecisionScopeColor=" << format_hex_color_string(*g_alpine_game_config.precision_scope_color_override) << "\n";
+    }
+    if (g_alpine_game_config.rail_scope_color_override) {
+        file << "RailScopeColor=" << format_hex_color_string(*g_alpine_game_config.rail_scope_color_override) << "\n";
+    }
+    if (g_alpine_game_config.ar_ammo_digit_color_override) {
+        file << "ArAmmoColor=" << format_hex_color_string(*g_alpine_game_config.ar_ammo_digit_color_override) << "\n";
+    }
+    if (g_alpine_game_config.damage_notify_color_override) {
+        file << "DamageNotifyColor=" << format_hex_color_string(*g_alpine_game_config.damage_notify_color_override) << "\n";
+    }
 
     // Singleplayer
     file << "\n[SingleplayerSettings]\n";