+++ docs/CHANGELOG.md
@@ -117,6 +117,8 @@ Version 1.2.0 (Willow): Not yet released
 - Alias `Ping Location` control to reset to freelook spectator camera accel to base value
 - Add scroll wheel support to message log panel and scrollable item lists in menus
 - Add `kick after warning` option to dedicated server client inactivity tracking
+- Add `ui_color_sniper_scope`, `ui_color_precision_scope`, and `ui_color_rail_scope` commands
+- Add `ui_color_ar_ammo` and `ui_color_damage_notify` commands
 
 [@is-this-c](https://github.com/is-this-c)
 - Add `Tlv{Writer, Reader}` and `HighResTimer`